requirements.txt
edge-tts>=6.1.0
gtts>=2.3.0
supabase>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
readability-lxml>=0.8.1
newspaper3k>=0.2.8
pydub>=0.25.1
flask>=2.3.0
flask-cors>=4.0.0